feat(LoginExpirationBanner): move in a dedicated feature -> loginFeatures
