+++ locales/en/index.yml
@@ -4184,10 +4184,12 @@ fastLogin:
   whatsNew:
     title: Logging into IO just got easier!
     subtitle: From now on, to log into the app you'll have to use your SPID or CIE only once a year. After that you can just use your unlock code, fingerprint or face.
-  expirationBanner:
-    content: Il {{date}} dovrai accedere di nuovo con SPID o CIE
-    action: 
-      label: Leggi come fare
+loginFeatures:
+  loginPreferences:
+    expirationBanner:
+      content: Il {{date}} dovrai accedere di nuovo con SPID o CIE
+      action: 
+        label: Leggi come fare
 whatsNew:
   title: What's changed?
 bonusCard:
