+++ locales/it/index.yml
@@ -4184,10 +4184,12 @@ fastLogin:
   whatsNew:
     title: Abbiamo migliorato l’accesso a IO!
     subtitle: Ora puoi accedere all'app utilizzando il tuo SPID o la tua CIE solo una volta all'anno. Ti basterà poi utilizzare il tuo codice di sblocco o il biometrico per entrare.
-  expirationBanner:
-    content: Il {{date}} dovrai accedere di nuovo con SPID o CIE
-    action: 
-      label: Leggi come fare
+loginFeatures:
+  loginPreferences:
+    expirationBanner:
+      content: Il {{date}} dovrai accedere di nuovo con SPID o CIE
+      action: 
+        label: Leggi come fare
 whatsNew:
   title: Cosa cambia?
 bonusCard:
