+++ ts/features/fastLogin/store/selectors/index.ts
@@ -1,13 +1,9 @@
 import { createSelector } from "reselect";
 import { uniqWith, isEqual } from "lodash";
-import * as O from "fp-ts/lib/Option";
-import { differenceInDays } from "date-fns";
-import { pipe } from "fp-ts/lib/function";
 import { remoteConfigSelector } from "../../../../store/reducers/backendStatus/remoteConfig";
 import { fastLoginOptIn, fastLoginEnabled } from "../../../../config";
 import { GlobalState } from "../../../../store/reducers/types";
 import { isPropertyWithMinAppVersionEnabled } from "../../../../store/reducers/featureFlagWithMinAppVersionStatus";
-import { sessionInfoSelector } from "../../../../store/reducers/authentication";
 
 export const fastLoginOptInSelector = (state: GlobalState) =>
   state.features.loginFeatures.fastLogin.optIn;
@@ -101,51 +97,3 @@ export const isFastLoginUserInteractionNeededForSessionExpiredSelector = (
 
 export const tokenRefreshSelector = (state: GlobalState) =>
   state.features.loginFeatures.fastLogin.tokenRefreshHandler.tokenRefresh;
-
-/**
- * This selector returns a boolean that indicates whether to show
- * the sessionExpirationBanner or not
- */
-const showSessionExpirationBannerSelector = (state: GlobalState) =>
-  state.features.loginFeatures.fastLogin.sessionExpiration
-    .showSessionExpirationBanner;
-
-/**
- * This selector combines control over the value of `expirationDate`,
- * ensuring that the difference with the actual date at the time of the check
- * is less than a specified value in days (e.g., 30),
- * with the value related to the `showSessionExpirationBanner` data.
- */
-export const isSessionExpirationBannerRenderableSelector = createSelector(
-  sessionInfoSelector,
-  remoteConfigSelector,
-  showSessionExpirationBannerSelector,
-  isFastLoginEnabledSelector,
-  (sessionInfo, config, showSessionExpirationBanner, isFastLogin) =>
-    pipe(
-      O.Do,
-      O.bind("expirationDate", () =>
-        pipe(
-          sessionInfo,
-          O.chainNullableK(({ expirationDate }) => expirationDate)
-        )
-      ),
-      O.bind("threshold", () =>
-        pipe(
-          config,
-          O.chainNullableK(({ loginConfig }) =>
-            isFastLogin
-              ? loginConfig?.notifyExpirationThreshold?.fastLogin
-              : loginConfig?.notifyExpirationThreshold?.standardLogin
-          )
-        )
-      ),
-      O.map(
-        ({ expirationDate, threshold }) =>
-          threshold !== 0 &&
-          differenceInDays(expirationDate, new Date()) < threshold &&
-          showSessionExpirationBanner
-      ),
-      O.getOrElse(() => false)
-    )
-);
