+++ ts/features/login/preferences/components/LoginExpirationBanner.tsx
@@ -5,11 +5,11 @@ import {
 } from "@pagopa/io-app-design-system";
 import { useCallback } from "react";
 import { StyleSheet, View } from "react-native";
-import { useIODispatch, useIOSelector } from "../../../store/hooks";
-import { openWebUrl } from "../../../utils/url";
-import I18n from "../../../i18n";
-import { closeSessionExpirationBanner } from "../store/actions/sessionExpirationActions";
-import { formattedExpirationDateSelector } from "../../../store/reducers/authentication";
+import { useIODispatch, useIOSelector } from "../../../../store/hooks";
+import { openWebUrl } from "../../../../utils/url";
+import I18n from "../../../../i18n";
+import { closeSessionExpirationBanner } from "../store/actions";
+import { formattedExpirationDateSelector } from "../../../../store/reducers/authentication";
 
 const SESSION_EXPIRED_EDUCATIONAL_URL =
   "https://assistenza.ioapp.it/hc/it/articles/32616176301713-Cosa-fare-quando-la-sessione-scade";
@@ -21,7 +21,7 @@ type Props = {
  * to use in case the banner's visibility has to be handled externally
  * (see MultiBanner feature for the landing screen)
  */
-export const FastLoginExpirationBanner = ({ handleOnClose }: Props) => {
+export const LoginExpirationBanner = ({ handleOnClose }: Props) => {
   const expirationDate = useIOSelector(formattedExpirationDateSelector);
   const { error } = useIOToast();
   const dispatch = useIODispatch();
@@ -39,10 +39,15 @@ export const FastLoginExpirationBanner = ({ handleOnClose }: Props) => {
   return (
     <View style={styles.margins}>
       <Banner
-        content={I18n.t("fastLogin.expirationBanner.content", {
-          date: expirationDate
-        })}
-        action={I18n.t("fastLogin.expirationBanner.action.label")}
+        content={I18n.t(
+          "loginFeatures.loginPreferences.expirationBanner.content",
+          {
+            date: expirationDate
+          }
+        )}
+        action={I18n.t(
+          "loginFeatures.loginPreferences.expirationBanner.action.label"
+        )}
         pictogramName="identityCheck"
         color="neutral"
         size="big"
