+++ ts/features/login/preferences/store/actions/index.ts
@@ -4,6 +4,6 @@ export const closeSessionExpirationBanner = createStandardAction(
   "CLOSE_SESSION_EXPIRATION_BANNER"
 )();
 
-export type FastLoginSessionExpirationActions = ActionType<
+export type LoginPreferencesActions = ActionType<
   typeof closeSessionExpirationBanner
 >;
