+++ ts/store/actions/types.ts
@@ -35,7 +35,7 @@ import { IngressScreenActions } from "../../features/ingress/store/actions";
 import { MixpanelFeatureActions } from "../../features/mixpanel/store/actions";
 import { LandingScreenBannerActions } from "../../features/landingScreenMultiBanner/store/actions";
 import { SpidConfigActions } from "../../features/spidLogin/store/actions";
-import { FastLoginSessionExpirationActions } from "../../features/fastLogin/store/actions/sessionExpirationActions";
+import { LoginPreferencesActions } from "../../features/login/preferences/store/actions";
 import { AnalyticsActions } from "./analytics";
 import { ApplicationActions } from "./application";
 import { AuthenticationActions } from "./authentication";
@@ -109,7 +109,7 @@ export type Action =
   | MixpanelFeatureActions
   | LandingScreenBannerActions
   | SpidConfigActions
-  | FastLoginSessionExpirationActions;
+  | LoginPreferencesActions;
 
 export type Dispatch = DispatchAPI<Action>;
